+++ cpp/src/Freeze/Util.h
@@ -34,7 +34,7 @@ inline void
 initializeInDbt(const std::vector<Ice::Byte>& v, Dbt& dbt)
 {
     dbt.set_data(const_cast<Ice::Byte*>(&v[0]));
-    dbt.set_size(v.size());
+    dbt.set_size(static_cast<u_int32_t>(v.size()));
     dbt.set_ulen(0);
     dbt.set_dlen(0);
     dbt.set_doff(0);
@@ -47,7 +47,7 @@ initializeOutDbt(std::vector<Ice::Byte>& v, Dbt& dbt)
     v.resize(v.capacity());
     dbt.set_data(&v[0]);
     dbt.set_size(0);
-    dbt.set_ulen(v.size());
+    dbt.set_ulen(static_cast<u_int32_t>(v.size()));
     dbt.set_dlen(0);
     dbt.set_doff(0);
     dbt.set_flags(DB_DBT_USERMEM);
